AUTHENTICATION WITH MIDDLEWARE COMPLETED

+++ server/controllers/goalsControllers.js
@@ -1,8 +1,9 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -14,6 +15,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
 
   res.status(200).json(goal);
@@ -25,6 +27,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  //Make sure the logged in user matches the goal user 
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error("User not authorized")
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -32,16 +46,29 @@ const updateGoal = asyncHandler(async (req, res) => {
 });
 
 const deleteGoal = asyncHandler(async (req, res) => {
-    const goal = await Goal.findById(req.params.id).lean();
-    if (!goal) {
-      res.status(400);
-      throw new Error("Goal not found");
-    }
-  
-    await Goal.deleteOne({ _id: goal._id });
-    res.status(200).json({ message: `Delete Goal from controllers ${req.params.id} ` });
-  });
-  
+  const goal = await Goal.findById(req.params.id).lean();
+  if (!goal) {
+    res.status(400);
+    throw new Error("Goal not found");
+  }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("user not found");
+  }
+
+  //Make sure the logged in user matches the goal user 
+  if(goal.user.toString() !== user.id){
+    res.status(401)
+    throw new Error("User not authorized")
+  }
+
+  await Goal.deleteOne({ _id: goal._id });
+  res
+    .status(200)
+    .json({ message: `Delete Goal from controllers ${req.params.id} ` });
+});
+
 module.exports = {
   getGoals,
   setGoal,
